Add code to calculate elasticity
The code in tex2elas is mostly taken from comitac_flow
with some slight modification to handle the fact that we
have rotation matrices in a single big numpy array and
in comitac_flow we haev a python list of arrays.

The change to theiaptfile adds a method to the particle
object that uses the rotation matrices and volume fraction
along with a (hard coded) single crystal constat to calculate
c_ij on the reference frame of the Earth.